fix(p4.1.1): CI verde — causas raiz de Linux #51 e Android #45
CI Linux #51 (única falha real, determinística): teste P1 readback codificava
a geometria ANTIGA do gizmo MOVE (30 vértices); o P4.1 (D3/D4) redesenhou a
affordance (4 setas + centro = 5 quads × 6 + 4 hastes × 6 = 54 — observado no
próprio CI). Contagem atualizada à especificação nova; provas de pixel e drag
intactas. NÃO é skip/enfraquecimento — o contrato de topologia do gizmo novo
fica pinado.

Os banners 'FAILED ... SIGSEGV' dos forenses P3.1/P3.5 no log são dos FILHOS
(fork) — decomposição dos summaries: pai 127|126|1 (só gizmo), filhos
121|119|2 e 127|125|2 (morte por sinal POR DESIGN, documentada no teste).
O pai passou nos dois jobs. Zero mudanças nos testes forenses.

CI Android #45 (clang NDK r27, -Werror; eng_audio morria antes de chegar a
EditorJni.cpp):
- OpenSlEsBackend: formatResult nunca existiu no TU (helper local com tabela
  Khronos); refuse/cleanupAndRefuse devolvem Result<void> via makeUnexpected
  (8 retornos); SL_IID_* = variáveis globais — dlsym dá o ENDEREÇO, valor
  vem da desreferência (null-check antes); SLEngineItf é duplo ponteiro —
  (*engine)->CreateOutputMix/CreateAudioPlayer.
- AutoBackend: createAAudioBackend agora declarada em Audio.hpp.
- EditorActivity.kt: 0x99000000 > Int.MAX_VALUE — .toInt() explícito.

Verificação local: linux-debug 31/31, linux-release 31/31, NDK r27
arm64-v8a+x86_64 230/230 alvos (libgoni.so linkada), gradlew assembleDebug
BUILD SUCCESSFUL. Adenda completa em docs/p4-editor-ux.md.

+++ android/app/src/main/java/com/goni/runtime/EditorActivity.kt
@@ -366,7 +366,9 @@ class EditorActivity : Activity(), SurfaceHolder.Callback2,
         playHud = TextView(this).apply {
             setTextColor(Ui.TEXT)
             setTextSize(TypedValue.COMPLEX_UNIT_SP, 11f)
-            setBackgroundColor(0x99000000)
+            // 0x99000000 > Int.MAX_VALUE — Kotlin exige .toInt() explícito
+            // (CI Android #45: literal não conforma a Int sem a conversão).
+            setBackgroundColor(0x99000000.toInt())
             setPadding(dp(8), dp(4), dp(8), dp(4))
             visibility = View.GONE
         }
